v 3.01

+++ pr1_ex3/Program.cs
@@ -1,3 +1,3 @@
-﻿Func<int, int, int, float> avg = (int x, int y, int z) => (x + y + z) / 3; 
+﻿Func<int, int, int, double> avg = (int x, int y, int z) => (x + y + z) / 3.0; 
 
 System.Console.WriteLine(avg(1, 1, 5));
